DelayedDatasetDisk

A class that contains a dataset where its samples are backed on disk. Each sample is stored on a file, and when queued operations are executed files are loaded and saved on a new folder.

Details

This class is not exported, but if you want to use it reach us at https://github.com/sipss/GCIMS/issues/ and we will export it.

Super class

GCIMS::DelayedDatasetBase -> DelayedDatasetDisk

Active bindings

sampleNames

The character vector with unique sample names. Renaming samples renames files on obj$getCurrentDir() as well

Methods

Public methods

DelayedDatasetDisk$new()

DelayedDatasetDisk$getSample()

DelayedDatasetDisk$getCurrentDir()

DelayedDatasetDisk$clone()

Inherited methods

GCIMS::DelayedDatasetBase$appendDelayedOp()

GCIMS::DelayedDatasetBase$hasDelayedOps()

GCIMS::DelayedDatasetBase$history_as_list()

GCIMS::DelayedDatasetBase$pending_as_list()

GCIMS::DelayedDatasetBase$realize()

GCIMS::DelayedDatasetBase$registerOptimization()

Method new()

Create a delayed dataset on disk

Usage

DelayedDatasetDisk$new(
  samples,
  scratch_dir,
  keep_intermediate = FALSE,
  dataset = NULL
)

Arguments

samples

A named vector. The names are sample ids, the values are either filenames or sample objects. If they are not filenames then the objects are dumped to disk. If they are filenames, the filenames are relative to base_dir.

scratch_dir

The directory where samples being processed will be saved

keep_intermediate

A logical value, whether intermediate realization steps should be saved.

dataset

The dataset that stores aggregated results from the processing

Returns

The DelayedDatasetDisk object

Method getSample()

Get a sample from the dataset

Usage

DelayedDatasetDisk$getSample(sample)

Arguments

sample

Either an integer (sample index) or a string (sample name)

Returns

The sample object

Method getCurrentDir()

Get the path to the location of the processed samples

Usage

DelayedDatasetDisk$getCurrentDir()

Returns

The full path to the directory where samples are saved

Method clone()

The objects of this class are cloneable with this method.

Usage

DelayedDatasetDisk$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.