Create a GCIMSChromatogram object

Create a GCIMSChromatogram object

GCIMSChromatogram(
  retention_time,
  intensity,
  drift_time_idx = NA_integer_,
  drift_time_ms = NA_real_,
  description = "",
  baseline = NULL,
  peaks = NULL,
  peaks_debug_info = NULL
)

Arguments

retention_time

A numeric vector with retention times

intensity

A numeric vector with the corresponding intensities

drift_time_idx

The index or indices used to get the intensity

drift_time_ms

The drift times corresponding to drift_time_idx.

description

A string with a description (used as plot title, useful e.g. to know the sample it came from)

baseline

A numeric vector of the same length as intensity with the corresponding baseline, or NULL if not set. Use estimateBaseline() to estimate it, baseline() to directly access it.

peaks

A data frame with peaks, use findPeaks() to fill it, or peaks() to set/get it

peaks_debug_info

A list with arbitrary debug information from findPeaks()

Value

A GCIMSChromatogram object

See also

Examples

GCIMSChromatogram(
  retention_time = seq(from = 0, to = 10, length.out = 200),
  intensity = 1:200
)
#> An object of class "GCIMSChromatogram"
#> Slot "retention_time":
#>   [1]  0.00000000  0.05025126  0.10050251  0.15075377  0.20100503  0.25125628
#>   [7]  0.30150754  0.35175879  0.40201005  0.45226131  0.50251256  0.55276382
#>  [13]  0.60301508  0.65326633  0.70351759  0.75376884  0.80402010  0.85427136
#>  [19]  0.90452261  0.95477387  1.00502513  1.05527638  1.10552764  1.15577889
#>  [25]  1.20603015  1.25628141  1.30653266  1.35678392  1.40703518  1.45728643
#>  [31]  1.50753769  1.55778894  1.60804020  1.65829146  1.70854271  1.75879397
#>  [37]  1.80904523  1.85929648  1.90954774  1.95979899  2.01005025  2.06030151
#>  [43]  2.11055276  2.16080402  2.21105528  2.26130653  2.31155779  2.36180905
#>  [49]  2.41206030  2.46231156  2.51256281  2.56281407  2.61306533  2.66331658
#>  [55]  2.71356784  2.76381910  2.81407035  2.86432161  2.91457286  2.96482412
#>  [61]  3.01507538  3.06532663  3.11557789  3.16582915  3.21608040  3.26633166
#>  [67]  3.31658291  3.36683417  3.41708543  3.46733668  3.51758794  3.56783920
#>  [73]  3.61809045  3.66834171  3.71859296  3.76884422  3.81909548  3.86934673
#>  [79]  3.91959799  3.96984925  4.02010050  4.07035176  4.12060302  4.17085427
#>  [85]  4.22110553  4.27135678  4.32160804  4.37185930  4.42211055  4.47236181
#>  [91]  4.52261307  4.57286432  4.62311558  4.67336683  4.72361809  4.77386935
#>  [97]  4.82412060  4.87437186  4.92462312  4.97487437  5.02512563  5.07537688
#> [103]  5.12562814  5.17587940  5.22613065  5.27638191  5.32663317  5.37688442
#> [109]  5.42713568  5.47738693  5.52763819  5.57788945  5.62814070  5.67839196
#> [115]  5.72864322  5.77889447  5.82914573  5.87939698  5.92964824  5.97989950
#> [121]  6.03015075  6.08040201  6.13065327  6.18090452  6.23115578  6.28140704
#> [127]  6.33165829  6.38190955  6.43216080  6.48241206  6.53266332  6.58291457
#> [133]  6.63316583  6.68341709  6.73366834  6.78391960  6.83417085  6.88442211
#> [139]  6.93467337  6.98492462  7.03517588  7.08542714  7.13567839  7.18592965
#> [145]  7.23618090  7.28643216  7.33668342  7.38693467  7.43718593  7.48743719
#> [151]  7.53768844  7.58793970  7.63819095  7.68844221  7.73869347  7.78894472
#> [157]  7.83919598  7.88944724  7.93969849  7.98994975  8.04020101  8.09045226
#> [163]  8.14070352  8.19095477  8.24120603  8.29145729  8.34170854  8.39195980
#> [169]  8.44221106  8.49246231  8.54271357  8.59296482  8.64321608  8.69346734
#> [175]  8.74371859  8.79396985  8.84422111  8.89447236  8.94472362  8.99497487
#> [181]  9.04522613  9.09547739  9.14572864  9.19597990  9.24623116  9.29648241
#> [187]  9.34673367  9.39698492  9.44723618  9.49748744  9.54773869  9.59798995
#> [193]  9.64824121  9.69849246  9.74874372  9.79899497  9.84924623  9.89949749
#> [199]  9.94974874 10.00000000
#> 
#> Slot "intensity":
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162
#> [163] 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [181] 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198
#> [199] 199 200
#> 
#> Slot "baseline":
#> NULL
#> 
#> Slot "drift_time_idx":
#> [1] NA
#> 
#> Slot "drift_time_ms":
#> [1] NA
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "peaks":
#> NULL
#> 
#> Slot "peaks_debug_info":
#> NULL
#>