GCIMSChromatogram class

GCIMSChromatogram is an S4 class to store a GCIMS Chromatogram It can be a single chromatogram or the aggregation of several chromatograms.

# S3 method for class 'GCIMSChromatogram'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

# S4 method for class 'GCIMSChromatogram'
description(object)

# S4 method for class 'GCIMSChromatogram,ANY'
description(object) <- value

# S4 method for class 'GCIMSChromatogram'
peaks(object)

# S4 method for class 'GCIMSChromatogram'
peaks(object) <- value

# S4 method for class 'GCIMSChromatogram,ANY'
plot(x, y, ...)

Arguments

x

A GCIMSChromatogram object to plot

row.names

NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional

logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.

...

additional arguments to be passed to or from methods.

object

A GCIMSChromatogram object

value

A data frame with the peak list

y

From the generic plot function, ignored for GCIMSChromatogram class objects

Value

The description of the chromatogram

The chromatogram object

A data frame with the peaks in the chromatogram

The GCIMSChromatogram object

A ggplot2 plot object

Functions

as.data.frame(GCIMSChromatogram): Coerce to data frame

description(GCIMSChromatogram): Get the description

description(object = GCIMSChromatogram) <- value: Set the description

peaks(GCIMSChromatogram): Get the peak list

peaks(GCIMSChromatogram) <- value: Set the peak list

plot(x = GCIMSChromatogram, y = ANY): plot method

Slots

retention_time

A numeric vector with retention times

intensity

A numeric vector with the corresponding intensities

baseline

A numeric vector of the same length as intensity with the corresponding baseline, or NULL if not set. Use estimateBaseline() to estimate it, baseline() to directly access it.

drift_time_idx

The index or indices used to get the intensity

drift_time_ms

The drift times corresponding to drift_time_idx.

description

A string with a description (used as plot title, useful e.g. to know the sample it came from)

peaks

A data frame with peaks, use findPeaks() to fill it, or peaks() to set/get it

peaks_debug_info

A list with arbitrary debug information from findPeaks()

See also