GCIMSDataset class

GCIMSDataset is an S4 class to store a dataset

GCIMSDataset(
  pData,
  base_dir,
  parser = "default",
  scratch_dir = tempfile("GCIMSDataset_tempdir_"),
  keep_intermediate = FALSE,
  on_ram = FALSE
)

GCIMSDataset_fromList(
  samples,
  pData = NULL,
  scratch_dir = tempfile("GCIMSDataset_tempdir_"),
  keep_intermediate = FALSE,
  on_ram = TRUE
)

# S4 method for GCIMSDataset
getSample(object, sample)

# S4 method for GCIMSDataset
getTIS(object)

# S4 method for GCIMSDataset
plotRIC(object, rt_range = NULL, sample = NULL)

# S4 method for GCIMSDataset
pData(object)

# S4 method for GCIMSDataset,ANY
pData(object) <- value

# S4 method for GCIMSDataset
peaks(object)

# S4 method for GCIMSDataset
peaks(object) <- value

Arguments

pData

A data frame with at least the SampleID and FileName columns.

base_dir

A directory containing the file names described in pData

parser

Either a string "default" or a function that takes a filename and returns a GCIMSSample object

scratch_dir

A directory to save intermediate results.

keep_intermediate

A logical. Whether to keep or not intermediate files when realizing the GCIMSDataset object

on_ram

A logical. If TRUE, samples are kept on RAM. This is faster as long as you have enough memory to keep all samples. If this is TRUE, then scratch_dir and keep_intermediate are ignored.

samples

A named list of GCIMSSample objects. names should match pData$SampleID

object

A GCIMSDataset object

sample

A number or a string with the sample index or name. If NULL, all samples are returned

rt_range

The minimum and maximum retention times to extract (length 2 vector)

value

The data frame with a peak list

Value

A GCIMSDataset object

A GCIMSDataset object

The corresponding GCIMSSample

A matrix with samples in rows and the drift time in columns

A plot

Details

The actual data is not stored in memory, but read/saved from/to files as needed, so the dataset object scales with large number of samples.

Methods (by generic)

getSample(GCIMSDataset): class

getTIS(GCIMSDataset): Get Total Ion Spectra matrix

plotRIC(GCIMSDataset): Plot Reverse Ion Chromatograms

Functions

GCIMSDataset(): Constructor method

GCIMSDataset_fromList(): Constructor method

pData(GCIMSDataset): Get the phenotype data

pData(object = GCIMSDataset) <- value: Set the phenotype data

peaks(GCIMSDataset): Get the peak list

peaks(GCIMSDataset) <- value: Set the peak list

Examples

# Create a new GCIMSDataset with methods::new()
dummy_obj <- methods::new(
  "GCIMSDataset",
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
# Create a new GCIMSDataset with the convenient constructor function:
dummy_obj <-GCIMSDataset(
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
# Create a new GCIMSDataset with the convenient constructor function:
sample1 <- GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
dummy_obj <- GCIMSDataset_fromList(
  pData = data.frame(SampleID = "Sample1", Sex = "female"),
  samples = list(Sample1 = sample1)
)