Methods for the GCIMSSample class

Methods for the GCIMSSample class

# S3 method for GCIMSSample
[(x, i, j, ...)

# S3 method for GCIMSSample
dim(x)

# S3 method for GCIMSSample
subset(x, dt_idx = NULL, rt_idx = NULL, dt_range = NULL, rt_range = NULL, ...)

# S4 method for GCIMSSample
getChromatogram(
  object,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL,
  aggregate = colSums
)

# S4 method for GCIMSSample
getSpectrum(
  object,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL,
  aggregate = rowSums
)

# S4 method for GCIMSSample
description(object)

# S4 method for GCIMSSample,ANY
description(object) <- value

# S4 method for GCIMSSample
peaks(object)

# S4 method for GCIMSSample
peaks(object) <- value

Arguments

x

A GCIMSSample object

i

index for drift time to subset

j

index for retention time to subset

...

ignored

dt_idx

A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)

rt_idx

A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)

dt_range

The minimum and maximum drift times to extract (length 2 vector)

rt_range

The minimum and maximum retention times to extract (length 2 vector)

object

A GCIMSSample object

aggregate

Function that takes the subsetted intensity matrix according to the region of interest and aggregates the retention times, returning a vector representing the spectrum intensity. rowSums by default.

value

A data frame with the peak list

Value

[: object x with features i and cells j

An integer vector with the number of rows and columns of the matrix

subset: A subsetted GCIMSSample object

Functions

[: Simple subsetter for GCIMSSample objects

dim(GCIMSSample): Dimension of the data matrix

subset(GCIMSSample): Subset a GCIMSSample object

getChromatogram(GCIMSSample): Get the extracted ion chromatogram

getSpectrum(GCIMSSample): Get IMS spectrum

description(GCIMSSample): Get the description

description(object = GCIMSSample) <- value: Set the description

peaks(GCIMSSample): Get the peak list

peaks(GCIMSSample) <- value: Set the peak list

See also

Examples

# `[' examples

obj <- GCIMSSample(drift_time=1:2, retention_time=1:3, data = matrix(1:6, nrow=2, ncol=3))
dim(obj)
#> [1] 2 3
x <- GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
getChromatogram(x)
#> An object of class "GCIMSChromatogram"
#> Slot "retention_time":
#> [1] 1 2 3
#> 
#> Slot "intensity":
#>  1  2  3 
#>  3  7 11 
#> 
#> Slot "baseline":
#> NULL
#> 
#> Slot "drift_time_idx":
#> [1] 1 2
#> 
#> Slot "drift_time_ms":
#> [1] 1 2
#> 
#> Slot "description":
#> character(0)
#> 
#> Slot "peaks":
#> NULL
#> 
#> Slot "peaks_debug_info":
#> NULL
#> 
# Take the maximum intensity in the region for each retention time:
sp1 <- getChromatogram(x, aggregate = function(x) apply(x, 2, max))
x <- GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
getSpectrum(x, rt_idx = 2)
#> An object of class "GCIMSSpectrum"
#> Slot "drift_time":
#> [1] 1 2
#> 
#> Slot "intensity":
#> 1 2 
#> 3 4 
#> 
#> Slot "baseline":
#> NULL
#> 
#> Slot "retention_time_idx":
#> [1] 2
#> 
#> Slot "retention_time_s":
#> [1] 2
#> 
#> Slot "description":
#> character(0)
#> 

# Take the maximum intensity in the region for each drift time:
sp1 <- getSpectrum(x, aggregate = function(x) apply(x, 1, max))