Index normalization

Users may want to specify:

dt_rt_range_normalization(
  dt = numeric(0L),
  rt = numeric(0L),
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL
)

Arguments

dt

A numeric vector with the drift time

rt

A numeric vector with the retention time

dt_range

The minimum and maximum drift times to extract (length 2 vector)

rt_range

The minimum and maximum retention times to extract (length 2 vector)

dt_idx

A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)

rt_idx

A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)

Value

A named list with class dt_rt_range_normalization. The elements of the list are:

dt_ms_min, dt_ms_max: Drift time in milliseconds (min and max). This is the same value if the user did want a single value

dt_idx_min, dt_idx_max: The indices corresponding to the requested range limits in drift time

dt_logical: A logical vector of length equal to the given drift time vector, with TRUE where the range is included and FALSE elsewhere. Useful for indexing.

rt_s_min, rt_s_max: Retention time in seconds (min and max). This is the same value if the user did want a single value

rt_idx_min, rt_idx_max: The indices corresponding to the requested range limits in retention time.

rt_logical: A logical vector of length equal to the given retention time vector, with TRUE where the range is included and FALSE elsewhere. Useful for indexing.

Details

A single retention/drift time to extract, in physical units or in indices

A range of retention/drift time to extract, in physical units or in indices

This function is internal to the GCIMS-package and it is meant to be used to normalize those specifications, returning an object that computes the rest so the caller function can easily get the indices or physical units it needs.

The behaviour is intuitive in corner cases, where the conversion from physical units to indices may have rounding issues:

If a single value is requested in physical units the closest one will be returned. If the requested single value is outside the valid range, then an error is given. A tolerance of one time step is accepted to prevent rounding issues.

If a range value is requested in physical units, all the points within the range are included. The interval is considered closed at both ends, for consistency with how indexes behave in R.